Conflicts:
	main.js
	public/index.html
	public/js/events.js
	public/js/explorations.js
	public/js/user.js
	public/js/visualiser.js
	public/style.css

+++ main.js
@@ -27,7 +27,7 @@ var bodyParser = require('body-parser');
 var fs = require('fs');
 var app = express();
 
-app.use(bodyParser());
+app.use(bodyParser.json());
 app.use(bodyParser.json({limit: '50mb'}));
 app.use(bodyParser.urlencoded({limit: '50mb', extended: true}));
 //app.use(express.json({limit: '50mb'}));
@@ -43,33 +43,35 @@ app.use(express.static(__dirname + '/public'));
 app.post('/postpath', function(req, res){
 	var timestamp = new Date();
 
-	var path = req.body.path_taken;
+	var path = req.body;
 
 	// create 'path' path if it doesn't already exist
 	var filePath = "public/data/path/";
 	ensureDirExists(filePath)
 
-	fs.writeFile(filePath + timestamp + "-savePath.json", path+"\n", function(err){
+	fs.writeFile(filePath + timestamp + "-savePath.json", JSON.stringify(path, null, 4), function(err){
 		if(err){ console.log(err); }
 	});
 });
 
 //post exploration on the map for loading
 app.post('/postExploration', function(req, res){
+
+	var exploration = req.body;
+	var name = exploration.user.fname;
 	//var timestamp = new Date();
-	var timestamp = req.body.timestamp;
-	var exploration = req.body.exploration;
-	var name = req.body.name;
-	// makes 'pathectory' for files if none exist.
+	var timestamp = exploration.startTimeStamp;
+	// makes directory for files if none exist.
 	var path = "public/data/user/";
 	ensureDirExists(path);
 	path += "User-"+name;
 	ensureDirExists(path);
 	path +="/Exploration/";
-	ensureDirExists(path)
-	fs.writeFile(path + timestamp + ".json", exploration+"\n", function(err){
+	ensureDirExists(path);
+	fs.writeFile(path +name+ timestamp + ".json", JSON.stringify(exploration, null, 4), function(err){
 		if(err){ console.log(err); }
 	});
+	console.log("wrote exploration file");
 });
 
 //post file to shared user folder
@@ -79,12 +81,12 @@ app.post('/shareExploration', function(req, res){
 	var file = req.body.file;
 	var to = req.body.to;
 	var from = req.body.from;
-	console.log("to: "+ to + " from: "+ from);
-	// makes 'directory' for files if none exist.
+	console.log("shared exploration to: "+ to + " from: "+ from);
+	// makes directory for files if none exist.
 	var path = "public/data/user/User-" + to+"/";
 	ensureDirExists(path);
 	ensureDirExists(path +"Shared/");
-	fs.writeFile(path +"Shared/" + from  +"-"+ timestamp + ".json", file +"\n", function(err){
+	fs.writeFile(path +"/Shared/" + from  +"-"+ timestamp + ".json", file +"\n", function(err){
 		if(err){
 			console.log(err);
 		}
@@ -93,7 +95,7 @@ app.post('/shareExploration', function(req, res){
 
 app.post('/postAnnotation', function(req, res){
 
-	var annotation = JSON.parse(req.body.annotation);
+	var annotation = req.body;
 	var timestamp = new Date(annotation.timestamp); // apparently timestamp is now a string...
 	var location = annotation.location;
 	var user = annotation.user; // string
@@ -112,7 +114,7 @@ app.post('/postAnnotation', function(req, res){
 	fs.writeFile(fileName, JSON.stringify(annotation, 4, null), function(err) {
 		if (err){ console.log("errooor: "+err); }
 	});
-	res.send(200); // success code
+	res.sendStatus(200); // success code
 });
 
 app.get("/getAnnotation", function(req, res){
@@ -173,7 +175,7 @@ app.get("/getNotification", function(req, res){
 
 app.post("/deleteAnnotation", function(req, res){
 	console.log("delete annotation");
-	var annotation = JSON.parse(req.body.annotation);
+	var annotation = req.body;
 	var locationName = annotation.location.properties.NAME;
 	var path = "public/data/annotation/" + locationName + "/";
 	var annotationFiles = fs.readdirSync(path);
@@ -184,11 +186,11 @@ app.post("/deleteAnnotation", function(req, res){
 
 		//console.log(JSON.stringify(annotation) + "\n" + JSON.stringify(inputAnnotation) + "\n\n");
 		// if annotations are equal, delete the file
-		if (//annotation.user === inputAnnotation.user
-				annotation.timestamp === inputAnnotation.timestamp
+		if (annotation.user.fname === inputAnnotation.user.fname
+				&&annotation.timestamp === inputAnnotation.timestamp
 				&& annotation.text === inputAnnotation.text){
 			fs.unlink(path + filename);
-			res.send(200); // success code
+			res.sendStatus(200); // success code
 		}
 	});
 });
@@ -214,6 +216,7 @@ app.post("/deletePlayed", function(req, res){
 	console.log("delete Played file from shared folder");
 	var playedFile = JSON.parse(req.body.playedFile);
 	var userName = playedFile.user.fname;
+
 	var path = "public/data/user/";
 	ensureDirExists(path);
 	path +="User-" + userName+ "/";
@@ -225,10 +228,13 @@ app.post("/deletePlayed", function(req, res){
 	// find and delete the file corresponding to the annotation specified.
 	sharedFiles.forEach(function(filename){
 
-		console.log("aa: "+path + filename);
+		//console.log("aa: "+path + filename);
 		var inputSharedFile = JSON.parse(fs.readFileSync(path + filename));
 		console.log(playedFile.startTimeStamp.localeCompare(inputSharedFile.startTimeStamp)==0);
+		console.log("playedFile.startTimeStamp: "+playedFile.startTimeStamp);
+		console.log("inputSharedFile.startTimeStamp: "+inputSharedFile.startTimeStamp);
 		if (playedFile.startTimeStamp.localeCompare(inputSharedFile.startTimeStamp)==0){
+
 			fs.unlink(path + filename);
 			res.send(200); // success code
 		}
